requirements.txt
requests
pywhatkit
schedule
python-dotenv
firebase-admin
